pyproject.toml
[project]
name = "promptlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
